addend end to end TCP specs
