+++ tests/TurboMqtt.Core.Tests/End2End/TcpMqtt311End2EndSpecs.cs
@@ -4,8 +4,6 @@
 // </copyright>
 // -----------------------------------------------------------------------
 
-using Akka.Hosting;
-using Akka.TestKit.Xunit2;
 using TurboMqtt.Core.Client;
 using TurboMqtt.Core.IO;
 using TurboMqtt.Core.Protocol;
@@ -13,24 +11,20 @@ using Xunit.Abstractions;
 
 namespace TurboMqtt.Core.Tests.End2End;
 
-public class TcpMqtt311End2EndSpecs : TestKit
+public class TcpMqtt311End2EndSpecs : TransportSpecBase
 {
     public TcpMqtt311End2EndSpecs(ITestOutputHelper output) : base(output: output)
     {
         _server = new FakeMqttTcpServer(new MqttTcpServerOptions("localhost", 21883), MqttProtocolVersion.V3_1_1, Log);
         _server.Bind();
-        ClientFactory = new MqttClientFactory(Sys);
     }
     
     private readonly FakeMqttTcpServer _server;
-    public MqttClientFactory ClientFactory { get; }
-    
-    public MqttClientConnectOptions DefaultConnectOptions =>
-        new MqttClientConnectOptions("test-client", MqttProtocolVersion.V3_1_1)
-        {
-            UserName = "test",
-            Password = "test",
-        };
+    public override async Task<IMqttClient> CreateClient()
+    {
+        var client = await ClientFactory.CreateTcpClient(DefaultConnectOptions, DefaultTcpOptions);
+        return client;
+    }
     
     public MqttClientTcpOptions DefaultTcpOptions => new("localhost", 21883);
 
@@ -39,18 +33,6 @@ public class TcpMqtt311End2EndSpecs : TestKit
         // shut down our local TCP server
         _server.Shutdown();
         base.AfterAll();
-    }
-    
-    [Fact]
-    public async Task ShouldConnectAndDisconnect()
-    {
-        var client = await ClientFactory.CreateTcpClient(DefaultConnectOptions, DefaultTcpOptions);
-
-        using var cts = new CancellationTokenSource(RemainingOrDefault);
-        var connectResult = await client.ConnectAsync(cts.Token);
-        connectResult.IsSuccess.Should().BeTrue();
-        await client.DisconnectAsync(cts.Token);
-        client.IsConnected.Should().BeFalse();
     }
 
     [Fact]
